Move to deployment.resource
