+++ harvester_robot_tests/libs/addon/rest.py
@@ -500,86 +500,6 @@ class Rest(Base):
             f"to be gone after {timeout}s"
         )
 
-    def wait_for_deployment_ready(self, name, namespace, timeout=DEFAULT_TIMEOUT):
-        """
-        Wait for a Deployment to report all replicas ready (REST)
-
-        Args:
-            name: Name of the deployment
-            namespace: Kubernetes namespace
-            timeout: Timeout in seconds
-
-        Returns:
-            bool: True once the deployment is ready
-        """
-        logging(f"Waiting for deployment '{namespace}/{name}' to be ready (REST)")
-        retry_count, retry_interval = get_retry_count_and_interval()
-        max_retries = int(timeout / retry_interval)
-
-        for i in range(max_retries):
-            try:
-                code, data = self.api_client.get(f"v1/apps.deployments/{namespace}/{name}")
-                if code == 200 and data:
-                    desired = data.get('spec', {}).get('replicas', 0) or 0
-                    ready = data.get('status', {}).get('readyReplicas', 0) or 0
-                    if desired > 0 and ready == desired:
-                        logging(
-                            f"Deployment '{namespace}/{name}' is ready ({ready}/{desired})"
-                        )
-                        return True
-                    logging(
-                        f"Deployment '{namespace}/{name}' not ready ({ready}/{desired}), "
-                        f"retrying... ({i+1}/{max_retries})"
-                    )
-                else:
-                    logging(
-                        f"Deployment '{namespace}/{name}' not available (HTTP {code}), "
-                        f"retrying... ({i+1}/{max_retries})"
-                    )
-            except Exception as e:
-                logging(f"Error reading deployment: {e}", level='WARNING')
-
-            time.sleep(retry_interval)
-
-        raise TimeoutError(
-            f"Timeout waiting for deployment '{namespace}/{name}' to be ready after {timeout}s"
-        )
-
-    def wait_for_deployment_gone(self, name, namespace, timeout=DEFAULT_TIMEOUT):
-        """
-        Wait for a Deployment to be removed (REST)
-
-        Args:
-            name: Name of the deployment
-            namespace: Kubernetes namespace
-            timeout: Timeout in seconds
-
-        Returns:
-            bool: True once the deployment no longer exists
-        """
-        logging(f"Waiting for deployment '{namespace}/{name}' to be gone (REST)")
-        retry_count, retry_interval = get_retry_count_and_interval()
-        max_retries = int(timeout / retry_interval)
-
-        for i in range(max_retries):
-            try:
-                code, _ = self.api_client.get(f"v1/apps.deployments/{namespace}/{name}")
-                if code == 404:
-                    logging(f"Deployment '{namespace}/{name}' is gone")
-                    return True
-                logging(
-                    f"Deployment '{namespace}/{name}' still present (HTTP {code}), "
-                    f"retrying... ({i+1}/{max_retries})"
-                )
-            except Exception as e:
-                logging(f"Error reading deployment: {e}", level='WARNING')
-
-            time.sleep(retry_interval)
-
-        raise TimeoutError(
-            f"Timeout waiting for deployment '{namespace}/{name}' to be gone after {timeout}s"
-        )
-
     def get_configmap_data(self, name, namespace):
         """
         Get the data map of a ConfigMap (REST)
